Add RestAuthenticator SPI, Authenticator interface, AuthResult enhancements, opt/opte refactor, and SonarLint fixes
- New RestAuthenticator SPI for resource-level authentication (@Rest(authenticator=...),
  AuthResultAccumulator, RestContext authentication chain, RestRequest.setAuthResult)
- New Authenticator shared interface; AuthFilter updated to implement it
- AuthResult enhancements: MergeMode, ofRoles(), replacing() factory methods
- New unit tests: AuthResultAccumulator_Test, AuthResult_Test, Authenticator_Test,
  RestAuthenticator_Test, and AuthFilterChain_Test additions
- New integration tests: RestAuthenticatorE2E_Test, RestContextAuthenticator_Test,
  RestRequestAuthResult_Test
- Docusaurus doc: 10.42.01.RestServerAuthenticator.md, sidebar update, release notes
- Codebase-wide opt/opte refactor (~244 replacements across ~58 files)
- SonarLint suppressions: AuthResult (S1192), ParquetParserSession (S107x3),
  RestRequest client (S3776), ProtobufParser (S115)
- SonarLint fixes: FluentStringAssertion regex (S5852), RrpcInterfaceMethodMeta
  array equals/hashCode/toString (S6218), ParquetParserSession computeIfAbsent (S3047)
- TODO comment removals: YamlParserSession, ParquetParserSession
- Utils_Test null-cast fix for generic type inference

+++ juneau-core/juneau-marshall/src/main/java/org/apache/juneau/marshall/parquet/ParquetColumnReader.java
@@ -158,7 +158,7 @@ final class ParquetColumnReader {
 		long nanosOfDay = readInt64();
 		int julianDay = readInt32();
 		// Julian day 2440588 == 1970-01-01.  Days since epoch * 86400s + nanos-of-day.
-		long epochDay = (long) julianDay - 2440588L;
+		long epochDay = julianDay - 2440588L;
 		long epochSecond = epochDay * 86_400L + nanosOfDay / 1_000_000_000L;
 		long nanoAdjust = nanosOfDay % 1_000_000_000L;
 		return Instant.ofEpochSecond(epochSecond, nanoAdjust);
